% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fc_rclust.R, R/fc_stable.R
\name{fc_rclust}
\alias{fc_rclust}
\title{Generate a List of Random kcca Objects.}
\usage{
fc_rclust(x, k, fc_cont, nrep = 100, verbose = FALSE, FUN = kcca,
  seed = 1234, plotme = TRUE)

fc_rclust(x, k, fc_cont, nrep = 100, verbose = FALSE, FUN = kcca,
  seed = 1234, plotme = TRUE)
}
\arguments{
\item{x}{Integer. matrix. Input to kcca.}

\item{k}{Integer. Number of clusters for this run.}

\item{nrep}{Integer. Number of repititions to run.}

\item{verbose}{Logical. Override for fc_control slot verbose.}

\item{FUN}{flexclust function.}

\item{seed}{Integer. Starting set.seed value for this run.}

\item{plotme}{Logical. Should plot be produced as side-effect?}

\item{fc_contol}{The flexclust control object for this run.}
}
\value{
A list(best, sizes, peak_at, tries)

tbl_df of k * nrep rows with cluster summary for k, seed, cluster #
}
\description{
For a given number of clusters, \code{k}, \code{nrep} kcca objects are generated.

Repeat flexclust runs
}
\details{
Each is kcca object is re-ordered so clusters are in decending size order. Cluster summary information is
pulled out of each object from the \code{clusinfo} slot. This includes the sizes of the clusters.

To look for stable cluster solutions, each random run is characterized by the sizes of the first
two clusters; which will be the largest after reording. \code{kde2d()} from the MASS package is
used to find density contours. The highest peak is determined and the distance of each solution
to the peak is retained for each cluster.

Optionally, the scatter plot of the sizes and corresponding coutour in plotted.
}

