% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fc_reorder.R
\name{fc_reorder}
\alias{fc_reorder}
\title{Reorder clusters in a kcca object.}
\usage{
fc_reorder(x, orderby = "decending size")
}
\arguments{
\item{x}{A kcca object.}

\item{orderby}{A string. Specifying the method to order by. Currently only "decending size".}
}
\value{
The kcca object with clusters reordered.
}
\description{
Since running kcca with different seeds will result, at least, in equivalent
clusters having different cluster sequence numbers which makes interpretation
of repeated runs difficult.
}
\details{
\code{fc_reorder} simply rearranges the clusters within the kcca object according
to the requested method.
}
\examples{
\dontrun{
fc_reorder(kcca(x, k, save.data = TRUE, control = fc_cont, family = kccaFamily(fc_family)))
}
}

